# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: "Mechanics Shop API"
  description: "A comprehensive REST API for managing automotive service operations including customers, service tickets, mechanics, and inventory"
  version: "1.0.0"
  contact:
    name: "API Support"
    email: "support@mechanicshop.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "http://127.0.0.1:5000"
    description: "Development server"
  - url: "https://api.mechanicshop.com"
    description: "Production server"
components:
  securitySchemes:
    customerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Customer JWT Token"
    mechanicAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Mechanic JWT Token"
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message describing what went wrong"
        details:
          type: object
          description: "Additional error details"
...
    MechanicLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "mechanic@example.com"
        password:
          type: string
          example: "mechanicpassword123"
    MechanicRegister:
      type: object
      required:
        - first_name
        - last_name
        - email
        - password
      properties:
        first_name:
          type: string
          example: "John"
          description: "Mechanic's first name"
        last_name:
          type: string
          example: "Doe"
          description: "Mechanic's last name"
        email:
          type: string
          format: email
          example: "john.doe@mechanicshop.com"
          description: "Mechanic's email address (must be unique)"
        password:
          type: string
          example: "securepassword123"
          description: "Mechanic's password"
    MechanicResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "john.doe@mechanicshop.com"

    # Service Ticket Schemas
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."

    # Health Check
...
paths:
...
  /mechanics/login:
    post:
      tags:
        - Authentication
      summary: "Mechanic Login"
      description: "Authenticate a mechanic and return a JWT token for accessing protected endpoints"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MechanicLogin'
            examples:
              validLogin:
                summary: "Valid Mechanic Login"
                value:
                  email: "mechanic@example.com"
                  password: "mechanicpassword123"
      responses:
        200:
          description: "Login successful"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        400:
          description: "Bad Request - Missing or invalid fields"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: "Unauthorized - Invalid credentials"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== CUSTOMER ENDPOINTS ==========
...
  /mechanics/register:
    post:
      tags:
        - Mechanics
      summary: "Register New Mechanic"
      description: "Create a new mechanic account with personal information and credentials"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MechanicRegister'
            examples:
              newMechanic:
                summary: "New Mechanic Registration"
                value:
                  first_name: "Jane"
                  last_name: "Smith"
                  email: "jane.smith@mechanicshop.com"
                  password: "securepassword123"
      responses:
        201:
          description: "Mechanic created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MechanicResponse'
        400:
          description: "Bad Request - Missing required fields"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        409:
          description: "Conflict - Email already registered"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /mechanics:
    get:
      tags:
        - Mechanics
      summary: "Get All Mechanics"
      description: "Retrieve all mechanics. Requires mechanic authentication."
      security:
        - mechanicAuth: []
      responses:
        200:
          description: "Mechanics retrieved successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MechanicResponse'
                  count:
                    type: integer
                    example: 5
        401:
          description: "Unauthorized - Missing or invalid token"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /mechanics/{id}:
    get:
      tags:
        - Mechanics
      summary: "Get Mechanic by ID"
      description: "Retrieve a specific mechanic by their ID. Requires mechanic authentication."
      security:
        - mechanicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: "Mechanic ID"
      responses:
        200:
          description: "Mechanic found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/MechanicResponse'
        404:
          description: "Mechanic not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Mechanics
      summary: "Update Mechanic"
      description: "Update mechanic information. Requires mechanic authentication."
      security:
        - mechanicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: "Mechanic ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MechanicRegister'
      responses:
        200:
          description: "Mechanic updated successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Mechanic updated successfully"
                  data:
                    $ref: '#/components/schemas/MechanicResponse'
        404:
          description: "Mechanic not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Mechanics
      summary: "Delete Mechanic"
      description: "Permanently delete a mechanic account. Requires mechanic authentication."
      security:
        - mechanicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: "Mechanic ID"
      responses:
        200:
          description: "Mechanic deleted successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Mechanic John Doe (ID: 1) deleted successfully"
        404:
          description: "Mechanic not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== SERVICE TICKET ENDPOINTS ==========
...